Re-doing how the items for the queue are allocated:
TargetQueueA holds items to grab
CountQueue holds the counts of those things to grab
TargetQueueB holds the store cells to place them

Each next thing to grab is processed to find how much can be held, if the store cell can hold it, and it will dynamically find more store cells to place any more as needed.

TODO: optimize so it doesn't search for items over and over
TODO: sort by thingdef so the store cells are accurate when hauling different things
TODO: don't hack in using the transpiler to the StoreUtilityCellSkipper. maybe just copy/paste and adapt the code.